✨ Feat(data): state manager
adding state manager with Jotai

FAI-63
