+++ apps/web/app/(app)/favorites/page.tsx
@@ -4,12 +4,12 @@ import { BookmarksList } from "@/components/bookmarks-list"
 import { filteredBookmarksAtom } from "@/lib/atoms"
 import { useAtomValue } from "jotai"
 
-export default function Page() {
+export default function Favorites() {
   const filteredBookmarks = useAtomValue(filteredBookmarksAtom)
 
   return (
     <BookmarksList
-      bookmarks={filteredBookmarks.filter((bookmark) => !bookmark.folderId)}
+      bookmarks={filteredBookmarks.filter((bookmark) => bookmark.favorite)}
     />
   )
 }
