--- apps/web/app/(app)/folder/[id]/page.tsx
@@ -1,11 +1,0 @@
-import { Button } from "@lucci/ui/components/button"
-
-export default function FolderContent() {
-  return (
-    <div className="flex min-h-svh items-center justify-center">
-      <div className="flex flex-col items-center justify-center gap-4">
-        <h1 className="text-2xl font-bold">Folder content</h1>
-      </div>
-    </div>
-  )
-}
