+++ apps/web/components/bookmarks-list.tsx
@@ -20,146 +20,11 @@ import { BookmarkPopover } from "./bookmark-popover"
 
 const cellStyle = "border-none"
 
-const bookmarks: Doc<"bookmarks">[] = [
-  {
-    _creationTime: Date.now() - 86400000, // 1 day ago
-    _id: "bookmark1" as any,
-    archived: false,
-    domain: "github.com",
-    favorite: true,
-    isPrivate: false,
-    metadata:
-      '{"title": "React Documentation", "description": "Official React documentation and tutorials", "image": "https://github.com/favicon.ico"}',
-    name: "React Documentation",
-    ownerId: "user1" as any,
-    tags: ["react", "frontend", "documentation"],
-    url: "https://react.dev",
-    workspaceId: "workspace1" as any,
-    folderId: "folder1-1-1" as any, // React folder
-    notes: "Essential React docs for the project",
-  },
-  {
-    _creationTime: Date.now() - 172800000, // 2 days ago
-    _id: "bookmark2" as any,
-    archived: false,
-    domain: "stackoverflow.com",
-    favorite: false,
-    isPrivate: false,
-    metadata:
-      '{"title": "How to use React hooks", "description": "Stack Overflow question about React hooks usage", "image": "https://stackoverflow.com/favicon.ico"}',
-    name: "React Hooks Best Practices",
-    ownerId: "user1" as any,
-    tags: ["react", "hooks", "stackoverflow"],
-    url: "https://stackoverflow.com/questions/react-hooks-best-practices",
-    workspaceId: "workspace1" as any,
-    folderId: "folder1-1-1-2" as any, // Hooks folder
-    notes: "Great examples of custom hooks",
-  },
-  {
-    _creationTime: Date.now() - 259200000, // 3 days ago
-    _id: "bookmark3" as any,
-    archived: false,
-    domain: "css-tricks.com",
-    favorite: true,
-    isPrivate: true,
-    metadata:
-      '{"title": "CSS Grid Layout", "description": "Complete guide to CSS Grid Layout", "image": "https://css-tricks.com/favicon.ico"}',
-    name: "CSS Grid Complete Guide",
-    ownerId: "user1" as any,
-    tags: ["css", "grid", "layout"],
-    url: "https://css-tricks.com/snippets/css/complete-guide-grid/",
-    workspaceId: "workspace1" as any,
-    folderId: "folder2-1" as any, // UI/UX folder
-    notes: "Personal reference for grid layouts",
-  },
-  {
-    _creationTime: Date.now() - 345600000, // 4 days ago
-    _id: "bookmark4" as any,
-    archived: false,
-    domain: "dev.to",
-    favorite: false,
-    isPrivate: false,
-    metadata:
-      '{"title": "TypeScript Tips", "description": "Advanced TypeScript tips and tricks", "image": "https://dev.to/favicon.ico"}',
-    name: "Advanced TypeScript Patterns",
-    ownerId: "user1" as any,
-    tags: ["typescript", "patterns", "advanced"],
-    url: "https://dev.to/typescript-advanced-patterns",
-    workspaceId: "workspace1" as any,
-    folderId: "folder1-1-1" as any, // React folder
-    notes: "Useful for complex type definitions",
-  },
-  {
-    _creationTime: Date.now() - 432000000, // 5 days ago
-    _id: "bookmark5" as any,
-    archived: true,
-    domain: "medium.com",
-    favorite: false,
-    isPrivate: false,
-    metadata:
-      '{"title": "Old React Patterns", "description": "Deprecated React patterns", "image": "https://medium.com/favicon.ico"}',
-    name: "Legacy React Class Components",
-    ownerId: "user1" as any,
-    tags: ["react", "legacy", "class-components"],
-    url: "https://medium.com/old-react-patterns",
-    workspaceId: "workspace1" as any,
-    folderId: "folder5" as any, // Old Stuff folder
-    notes: "For reference only - deprecated",
-  },
-  {
-    _creationTime: Date.now() - 518400000, // 6 days ago
-    _id: "bookmark6" as any,
-    archived: false,
-    domain: "nextjs.org",
-    favorite: true,
-    isPrivate: false,
-    metadata:
-      '{"title": "Next.js App Router", "description": "Next.js 13+ App Router documentation", "image": "https://nextjs.org/favicon.ico"}',
-    name: "Next.js App Router Guide",
-    ownerId: "user1" as any,
-    tags: ["nextjs", "app-router", "react"],
-    url: "https://nextjs.org/docs/app",
-    workspaceId: "workspace1" as any,
-    folderId: "folder1-1" as any, // Frontend folder
-    notes: "Essential for the new project",
-  },
-  {
-    _creationTime: Date.now() - 604800000, // 7 days ago
-    _id: "bookmark7" as any,
-    archived: false,
-    domain: "tailwindcss.com",
-    favorite: false,
-    isPrivate: false,
-    metadata:
-      '{"title": "Tailwind CSS", "description": "Utility-first CSS framework", "image": "https://tailwindcss.com/favicon.ico"}',
-    name: "Tailwind CSS Documentation",
-    ownerId: "user1" as any,
-    tags: ["css", "tailwind", "utility-first"],
-    url: "https://tailwindcss.com/docs",
-    workspaceId: "workspace1" as any,
-    folderId: "folder1-1-1-1" as any, // Components folder
-    notes: "For styling components",
-  },
-  {
-    _creationTime: Date.now() - 691200000, // 8 days ago
-    _id: "bookmark8" as any,
-    archived: false,
-    domain: "figma.com",
-    favorite: true,
-    isPrivate: true,
-    metadata:
-      '{"title": "Design System", "description": "Company design system in Figma", "image": "https://figma.com/favicon.ico"}',
-    name: "Company Design System",
-    ownerId: "user1" as any,
-    tags: ["design", "figma", "system"],
-    url: "https://figma.com/company-design-system",
-    workspaceId: "workspace1" as any,
-    folderId: "folder2-1-1" as any, // Design Systems folder
-    notes: "Internal design system reference",
-  },
-]
+interface Props {
+  bookmarks: Doc<"bookmarks">[]
+}
 
-export function BookmarksList() {
+export function BookmarksList({ bookmarks }: Props) {
   const formatDate = (timestamp: number) => {
     const date = new Date(timestamp)
     const now = new Date()
