+++ apps/web/components/nav-workspace.tsx
@@ -14,47 +14,13 @@ import {
   DropdownMenuTrigger,
 } from "@lucci/ui/components/dropdown-menu"
 import { ChevronsUpDown } from "lucide-react"
-import { Doc } from "@lucci/convex/generated/dataModel.js"
 import { WorkspaceCard } from "./workspace-card"
-
-const workspaces: Doc<"workspaces">[] = [
-  {
-    _creationTime: Date.now() - 86400000,
-    _id: "workspace1" as any,
-    icon: "🍕",
-    background: "bg-emerald-500",
-    members: [],
-    name: "Rigo's Workspace",
-    ownerId: "" as any,
-    userDefault: true,
-    notes: "",
-  },
-  {
-    _creationTime: Date.now() - 86400000,
-    _id: "workspace2" as any,
-    members: ["ssss" as any, "asddde"],
-    icon: "💸",
-    background: "bg-blue-500",
-    name: "Caban Energy",
-    ownerId: "" as any,
-    userDefault: true,
-    notes: "",
-  },
-  {
-    _creationTime: Date.now() - 86400000,
-    _id: "workspace3" as any,
-    members: [],
-    icon: "🚀",
-    background: "bg-purple-500",
-    name: "UCA",
-    ownerId: "" as any,
-    userDefault: true,
-    notes: "",
-  },
-]
+import { useAtomValue } from "jotai"
+import { selectedWorkspaceAtom, workspacesAtom } from "@/lib/atoms"
 
 export function NavWorkspace() {
-  const workspace = workspaces[0]
+  const workspaces = useAtomValue(workspacesAtom)
+  const workspace = useAtomValue(selectedWorkspaceAtom)
   const { isMobile } = useSidebar()
 
   return (
@@ -66,7 +32,7 @@ export function NavWorkspace() {
               size="lg"
               className="data-[state=open]:bg-sidebar-accent data-[state=open]:text-sidebar-accent-foreground"
             >
-              <WorkspaceCard workspace={workspace!} />
+              <WorkspaceCard workspace={workspace} />
               <ChevronsUpDown className="ml-auto size-4" />
             </SidebarMenuButton>
           </DropdownMenuTrigger>
@@ -77,7 +43,7 @@ export function NavWorkspace() {
             sideOffset={4}
           >
             <DropdownMenuLabel className="p-0 px-2 font-normal">
-              <WorkspaceCard workspace={workspace!} />
+              <WorkspaceCard workspace={workspace} />
             </DropdownMenuLabel>
             <DropdownMenuSeparator />
             <DropdownMenuGroup>
