+++ apps/web/components/navbar.tsx
@@ -1,3 +1,6 @@
+"use client"
+
+import { searchValueAtom } from "@/lib/atoms"
 import {
   Breadcrumb,
   BreadcrumbItem,
@@ -8,8 +11,11 @@ import {
 } from "@lucci/ui/components/breadcrumb"
 import { Input } from "@lucci/ui/components/input"
 import { SidebarTrigger } from "@lucci/ui/components/sidebar"
+import { useAtom } from "jotai"
 
 export function Navbar() {
+  const [searchValue, setSearchValue] = useAtom(searchValueAtom)
+
   return (
     <nav className="bg-background/50 absolute flex w-full items-center justify-between p-4 backdrop-blur-md">
       <div className="flex items-center gap-2">
@@ -32,7 +38,11 @@ export function Navbar() {
       </div>
 
       <div>
-        <Input placeholder="Search..." />
+        <Input
+          placeholder="Search..."
+          value={searchValue}
+          onChange={(e) => setSearchValue(e.target.value)}
+        />
       </div>
     </nav>
   )
