+++ apps/web/mocks/index.ts
@@ -1,0 +1,449 @@
+import { Doc } from "@lucci/convex/generated/dataModel.js"
+
+export const foldersMock: Doc<"folders">[] = [
+  {
+    isArchived: false,
+    isPrivate: false,
+    name: "Work",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: undefined,
+    _id: "folder1" as any,
+    _creationTime: Date.now() - 100000,
+  },
+  {
+    isArchived: false,
+    isPrivate: false,
+    name: "Frontend",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: "folder1" as any,
+    _id: "folder1-1" as any,
+    _creationTime: Date.now() - 95000,
+  },
+  {
+    isArchived: false,
+    isPrivate: false,
+    name: "Backend",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: "folder1" as any,
+    _id: "folder1-2" as any,
+    _creationTime: Date.now() - 90000,
+  },
+  {
+    isArchived: false,
+    isPrivate: false,
+    name: "Tools",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: "folder1" as any,
+    _id: "folder1-3" as any,
+    _creationTime: Date.now() - 85000,
+  },
+  {
+    isArchived: false,
+    isPrivate: true,
+    name: "Personal",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: undefined,
+    _id: "folder2" as any,
+    _creationTime: Date.now() - 90000,
+  },
+  {
+    isArchived: false,
+    isPrivate: true,
+    name: "UI/UX",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: "folder2" as any,
+    _id: "folder2-1" as any,
+    _creationTime: Date.now() - 85000,
+  },
+  {
+    isArchived: false,
+    isPrivate: true,
+    name: "Resources",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: "folder2" as any,
+    _id: "folder2-2" as any,
+    _creationTime: Date.now() - 80000,
+  },
+  {
+    isArchived: false,
+    isPrivate: true,
+    name: "Inspiration",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: "folder2" as any,
+    _id: "folder2-3" as any,
+    _creationTime: Date.now() - 75000,
+  },
+  {
+    isArchived: false,
+    isPrivate: false,
+    name: "Projects",
+    ownerId: "user2" as any,
+    workspaceId: "workspace2" as any,
+    parentFolderId: undefined,
+    _id: "folder3" as any,
+    _creationTime: Date.now() - 80000,
+  },
+  {
+    isArchived: false,
+    isPrivate: false,
+    name: "Articles",
+    ownerId: "user2" as any,
+    workspaceId: "workspace2" as any,
+    parentFolderId: "folder3" as any,
+    _id: "folder3-1" as any,
+    _creationTime: Date.now() - 75000,
+  },
+  {
+    isArchived: false,
+    isPrivate: false,
+    name: "Documentation",
+    ownerId: "user2" as any,
+    workspaceId: "workspace2" as any,
+    parentFolderId: "folder3" as any,
+    _id: "folder3-2" as any,
+    _creationTime: Date.now() - 70000,
+  },
+  {
+    isArchived: false,
+    isPrivate: false,
+    name: "2024",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: "folder1" as any,
+    _id: "folder4" as any,
+    _creationTime: Date.now() - 70000,
+  },
+  {
+    isArchived: true,
+    isPrivate: false,
+    name: "Old Stuff",
+    ownerId: "user2" as any,
+    workspaceId: "workspace2" as any,
+    parentFolderId: "folder3" as any,
+    _id: "folder5" as any,
+    _creationTime: Date.now() - 60000,
+  },
+  {
+    isArchived: false,
+    isPrivate: false,
+    name: "React",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: "folder1-1" as any,
+    _id: "folder1-1-1" as any,
+    _creationTime: Date.now() - 94000,
+  },
+  {
+    isArchived: false,
+    isPrivate: false,
+    name: "Components",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: "folder1-1-1" as any,
+    _id: "folder1-1-1-1" as any,
+    _creationTime: Date.now() - 93000,
+  },
+  {
+    isArchived: false,
+    isPrivate: false,
+    name: "Hooks",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: "folder1-1-1" as any,
+    _id: "folder1-1-1-2" as any,
+    _creationTime: Date.now() - 92000,
+  },
+  {
+    isArchived: false,
+    isPrivate: false,
+    name: "Vue",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: "folder1-1" as any,
+    _id: "folder1-1-2" as any,
+    _creationTime: Date.now() - 91000,
+  },
+  {
+    isArchived: false,
+    isPrivate: false,
+    name: "API",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: "folder1-2" as any,
+    _id: "folder1-2-1" as any,
+    _creationTime: Date.now() - 89000,
+  },
+  {
+    isArchived: false,
+    isPrivate: false,
+    name: "Endpoints",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: "folder1-2-1" as any,
+    _id: "folder1-2-1-1" as any,
+    _creationTime: Date.now() - 88000,
+  },
+  {
+    isArchived: false,
+    isPrivate: false,
+    name: "Database",
+    ownerId: "user1" as any,
+    workspaceId: "workspace1" as any,
+    parentFolderId: "folder1-2" as any,
+    _id: "folder1-2-2" as any,
+    _creationTime: Date.now() - 87000,
+  },
+]
+
+export const bookmarksMock: Doc<"bookmarks">[] = [
+  {
+    _creationTime: Date.now() - 86400000, // 1 day ago
+    _id: "bookmark1" as any,
+    archived: false,
+    domain: "github.com",
+    favorite: true,
+    isPrivate: false,
+    metadata:
+      '{"title": "React Documentation", "description": "Official React documentation and tutorials", "image": "https://github.com/favicon.ico"}',
+    name: "React Documentation",
+    ownerId: "user1" as any,
+    tags: ["react", "frontend", "documentation"],
+    url: "https://react.dev",
+    workspaceId: "workspace1" as any,
+    folderId: "folder1-1-1" as any, // React folder
+    notes: "Essential React docs for the project",
+  },
+  {
+    _creationTime: Date.now() - 172800000, // 2 days ago
+    _id: "bookmark2" as any,
+    archived: false,
+    domain: "stackoverflow.com",
+    favorite: false,
+    isPrivate: false,
+    metadata:
+      '{"title": "How to use React hooks", "description": "Stack Overflow question about React hooks usage", "image": "https://stackoverflow.com/favicon.ico"}',
+    name: "React Hooks Best Practices",
+    ownerId: "user1" as any,
+    tags: ["react", "hooks", "stackoverflow"],
+    url: "https://stackoverflow.com/questions/react-hooks-best-practices",
+    workspaceId: "workspace1" as any,
+    folderId: "folder1-1-1-2" as any, // Hooks folder
+    notes: "Great examples of custom hooks",
+  },
+  {
+    _creationTime: Date.now() - 259200000, // 3 days ago
+    _id: "bookmark3" as any,
+    archived: false,
+    domain: "css-tricks.com",
+    favorite: true,
+    isPrivate: true,
+    metadata:
+      '{"title": "CSS Grid Layout", "description": "Complete guide to CSS Grid Layout", "image": "https://css-tricks.com/favicon.ico"}',
+    name: "CSS Grid Complete Guide",
+    ownerId: "user1" as any,
+    tags: ["css", "grid", "layout"],
+    url: "https://css-tricks.com/snippets/css/complete-guide-grid/",
+    workspaceId: "workspace1" as any,
+    folderId: "folder2-1" as any, // UI/UX folder
+    notes: "Personal reference for grid layouts",
+  },
+  {
+    _creationTime: Date.now() - 345600000, // 4 days ago
+    _id: "bookmark4" as any,
+    archived: false,
+    domain: "dev.to",
+    favorite: false,
+    isPrivate: false,
+    metadata:
+      '{"title": "TypeScript Tips", "description": "Advanced TypeScript tips and tricks", "image": "https://dev.to/favicon.ico"}',
+    name: "Advanced TypeScript Patterns",
+    ownerId: "user1" as any,
+    tags: ["typescript", "patterns", "advanced"],
+    url: "https://dev.to/typescript-advanced-patterns",
+    workspaceId: "workspace1" as any,
+    folderId: "folder1-1-1" as any, // React folder
+    notes: "Useful for complex type definitions",
+  },
+  {
+    _creationTime: Date.now() - 432000000, // 5 days ago
+    _id: "bookmark5" as any,
+    archived: true,
+    domain: "medium.com",
+    favorite: false,
+    isPrivate: false,
+    metadata:
+      '{"title": "Old React Patterns", "description": "Deprecated React patterns", "image": "https://medium.com/favicon.ico"}',
+    name: "Legacy React Class Components",
+    ownerId: "user1" as any,
+    tags: ["react", "legacy", "class-components"],
+    url: "https://medium.com/old-react-patterns",
+    workspaceId: "workspace1" as any,
+    folderId: "folder5" as any, // Old Stuff folder
+    notes: "For reference only - deprecated",
+  },
+  {
+    _creationTime: Date.now() - 518400000, // 6 days ago
+    _id: "bookmark6" as any,
+    archived: false,
+    domain: "nextjs.org",
+    favorite: true,
+    isPrivate: false,
+    metadata:
+      '{"title": "Next.js App Router", "description": "Next.js 13+ App Router documentation", "image": "https://nextjs.org/favicon.ico"}',
+    name: "Next.js App Router Guide",
+    ownerId: "user1" as any,
+    tags: ["nextjs", "app-router", "react"],
+    url: "https://nextjs.org/docs/app",
+    workspaceId: "workspace1" as any,
+    folderId: "folder1-1" as any, // Frontend folder
+    notes: "Essential for the new project",
+  },
+  {
+    _creationTime: Date.now() - 604800000, // 7 days ago
+    _id: "bookmark7" as any,
+    archived: false,
+    domain: "tailwindcss.com",
+    favorite: false,
+    isPrivate: false,
+    metadata:
+      '{"title": "Tailwind CSS", "description": "Utility-first CSS framework", "image": "https://tailwindcss.com/favicon.ico"}',
+    name: "Tailwind CSS Documentation",
+    ownerId: "user1" as any,
+    tags: ["css", "tailwind", "utility-first"],
+    url: "https://tailwindcss.com/docs",
+    workspaceId: "workspace1" as any,
+    folderId: "folder1-1-1-1" as any, // Components folder
+    notes: "For styling components",
+  },
+  {
+    _creationTime: Date.now() - 691200000, // 8 days ago
+    _id: "bookmark8" as any,
+    archived: false,
+    domain: "figma.com",
+    favorite: true,
+    isPrivate: true,
+    metadata:
+      '{"title": "Design System", "description": "Company design system in Figma", "image": "https://figma.com/favicon.ico"}',
+    name: "Company Design System",
+    ownerId: "user1" as any,
+    tags: ["design", "figma", "system"],
+    url: "https://figma.com/company-design-system",
+    workspaceId: "workspace1" as any,
+    folderId: "folder2-1-1" as any, // Design Systems folder
+    notes: "Internal design system reference",
+  },
+
+  // Additional mocks with folderId: null (keeps all existing entries unchanged)
+  {
+    _creationTime: Date.now() - 7200000, // 2 hours ago
+    _id: "bookmark9" as any,
+    archived: false,
+    domain: "news.ycombinator.com",
+    favorite: false,
+    isPrivate: false,
+    metadata:
+      '{"title":"Hacker News","description":"Tech news aggregator","image":"https://news.ycombinator.com/favicon.ico"}',
+    name: "Hacker News Front Page",
+    ownerId: "user2" as any,
+    tags: ["news", "tech"],
+    url: "https://news.ycombinator.com",
+    workspaceId: "workspace2" as any,
+    folderId: null,
+    notes: "Daily tech headlines",
+  },
+  {
+    _creationTime: Date.now() - 3600000 * 6, // 6 hours ago
+    _id: "bookmark10" as any,
+    archived: false,
+    domain: "docs.google.com",
+    favorite: true,
+    isPrivate: true,
+    metadata:
+      '{"title":"Project Plan","description":"Shared project planning doc","image":"https://docs.google.com/favicon.ico"}',
+    name: "Q3 Project Plan",
+    ownerId: "user3" as any,
+    tags: ["planning", "internal"],
+    url: "https://docs.google.com/document/d/example",
+    workspaceId: "workspace1" as any,
+    folderId: null,
+    notes: "Keep updated with milestones",
+  },
+  {
+    _creationTime: Date.now() - 3600000 * 24 * 10, // 10 days ago
+    _id: "bookmark11" as any,
+    archived: false,
+    domain: "mdn.mozilla.org",
+    favorite: false,
+    isPrivate: false,
+    metadata:
+      '{"title":"MDN Fetch API","description":"Fetch API documentation","image":"https://developer.mozilla.org/favicon.ico"}',
+    name: "MDN - Fetch API",
+    ownerId: "user1" as any,
+    tags: ["web", "api", "reference"],
+    url: "https://developer.mozilla.org/en-US/docs/Web/API/Fetch_API",
+    workspaceId: "workspace1" as any,
+    folderId: null,
+    notes: "Reference for network calls",
+  },
+  {
+    _creationTime: Date.now() - 3600000 * 24 * 30, // 30 days ago
+    _id: "bookmark12" as any,
+    archived: true,
+    domain: "youtube.com",
+    favorite: false,
+    isPrivate: false,
+    metadata:
+      '{"title":"Conference Talk","description":"Frontend conference talk recording","image":"https://youtube.com/favicon.ico"}',
+    name: "Frontend Conf 2024 Talk",
+    ownerId: "user2" as any,
+    tags: ["video", "conference"],
+    url: "https://youtube.com/watch?v=example",
+    workspaceId: "workspace3" as any,
+    folderId: null,
+    notes: "Archived talk for reference",
+  },
+]
+
+export const workspacesMock: Doc<"workspaces">[] = [
+  {
+    _creationTime: Date.now() - 86400000,
+    _id: "workspace1" as any,
+    icon: "🍕",
+    background: "bg-emerald-500",
+    members: [],
+    name: "Rigo's Workspace",
+    ownerId: "" as any,
+    userDefault: true,
+    notes: "",
+  },
+  {
+    _creationTime: Date.now() - 86400000,
+    _id: "workspace2" as any,
+    members: ["ssss" as any, "asddde"],
+    icon: "💸",
+    background: "bg-blue-500",
+    name: "Caban Energy",
+    ownerId: "" as any,
+    userDefault: true,
+    notes: "",
+  },
+  {
+    _creationTime: Date.now() - 86400000,
+    _id: "workspace3" as any,
+    members: [],
+    icon: "🚀",
+    background: "bg-purple-500",
+    name: "UCA",
+    ownerId: "" as any,
+    userDefault: true,
+    notes: "",
+  },
+]
