+++ apps/web/package.json
@@ -13,8 +13,9 @@
   },
   "dependencies": {
     "@clerk/nextjs": "^6.31.4",
-    "@lucci/ui": "workspace:*",
     "@lucci/convex": "workspace:*",
+    "@lucci/ui": "workspace:*",
+    "jotai": "^2.13.1",
     "lucide-react": "^0.475.0",
     "next": "^15.4.5",
     "next-themes": "^0.4.6",
