+++ packages/convex/convex/_generated/dataModel.d.ts
@@ -13,14 +13,14 @@ import type {
   DocumentByName,
   TableNamesInDataModel,
   SystemTableNames,
-} from "convex/server";
-import type { GenericId } from "convex/values";
-import schema from "../schema.js";
+} from "convex/server"
+import type { GenericId } from "convex/values"
+import schema from "../schema.js"
 
 /**
  * The names of all of your Convex tables.
  */
-export type TableNames = TableNamesInDataModel<DataModel>;
+export type TableNames = TableNamesInDataModel<DataModel>
 
 /**
  * The type of a document stored in Convex.
@@ -30,7 +30,7 @@ export type TableNames = TableNamesInDataModel<DataModel>;
 export type Doc<TableName extends TableNames> = DocumentByName<
   DataModel,
   TableName
->;
+>
 
 /**
  * An identifier for a document in Convex.
@@ -46,7 +46,7 @@ export type Doc<TableName extends TableNames> = DocumentByName<
  * @typeParam TableName - A string literal type of the table name (like "users").
  */
 export type Id<TableName extends TableNames | SystemTableNames> =
-  GenericId<TableName>;
+  GenericId<TableName>
 
 /**
  * A type describing your Convex data model.
@@ -57,4 +57,4 @@ export type Id<TableName extends TableNames | SystemTableNames> =
  * This type is used to parameterize methods like `queryGeneric` and
  * `mutationGeneric` to make them type-safe.
  */
-export type DataModel = DataModelFromSchemaDefinition<typeof schema>;
+export type DataModel = DataModelFromSchemaDefinition<typeof schema>
